refresh files and change package name to com.example
