+++ app/src/androidTest/java/jp/wings/nikkeibp/getallsensors/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package jp.wings.nikkeibp.getallsensors
+package com.example.getallsensors
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("jp.wings.nikkeibp.getallsensors", appContext.packageName)
+        assertEquals("com.example.getallsensors", appContext.packageName)
     }
 }
