+++ app/src/main/java/com/example/getallsensors/MainActivity.kt
@@ -1,4 +1,4 @@
-package jp.wings.nikkeibp.getallsensors
+package com.example.getallsensors
 
 import android.content.Context
 import android.hardware.Sensor
