+++ app/src/main/java/com/example/getallsensors/SensorObject.kt
@@ -1,0 +1,7 @@
+package com.example.getallsensors
+
+data class SensorObject(
+    var name: String? = null,
+    var power: Float? = null,
+    var mindelay : Int? = null
+)
