--- app/src/main/java/jp/wings/nikkeibp/getallsensors/SensorObject.kt
@@ -1,7 +1,0 @@
-package jp.wings.nikkeibp.getallsensors
-
-data class SensorObject(
-    var name: String? = null,
-    var power: Float? = null,
-    var mindelay : Int? = null
-)
