+++ app/src/test/java/jp/wings/nikkeibp/getallsensors/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package jp.wings.nikkeibp.getallsensors
+package com.example.getallsensors
 
 import org.junit.Test
